+++ src/ast.c
@@ -93,6 +93,7 @@ Expr *expr_llamada(Arena *a, Expr *callee, Expr **args, int n_args, int linea, i
         e->como.llamada.callee = callee;
         e->como.llamada.args = args;
         e->como.llamada.n_args = n_args;
+        e->como.llamada.args_spread = NULL;  /* v1.22: NULL = sin spreads */
     }
     return e;
 }