+++ src/debug.c
@@ -163,6 +163,9 @@ int desensamblar_instruccion(const Chunk *c, int offset, FILE *out) {
         case OP_LLAMAR_CLASE:    return instruccion_byte("OP_LLAMAR_CLASE", c, offset, out);
         case OP_LLAMAR_METODO_LIGADO:
             return instruccion_byte("OP_LLAMAR_METODO_LIGADO", c, offset, out);
+        case OP_LISTA_AGREGAR:   return instruccion_simple("OP_LISTA_AGREGAR", offset, out);
+        case OP_LISTA_EXTENDER:  return instruccion_simple("OP_LISTA_EXTENDER", offset, out);
+        case OP_LLAMAR_SPREAD:   return instruccion_simple("OP_LLAMAR_SPREAD", offset, out);
         case OP_CLOSURE: {
             /* Formato: [byte fn_idx] [n_upvalues * (is_local, index)].
                No sabemos n_upvalues sin leer la FuncionBC del pool, así